refactor shutdown param fix tests enable android running in runner
